Moved them MUI object into helper file
