+++ src/App.js
@@ -3,6 +3,7 @@ import { BrowserRouter as Router, Route } from "react-router-dom"
 import { useAuth0 } from "./components/Auth/react-auth0-spa"
 import { connect } from "react-redux"
 import { loginUser } from "./actions"
+import theme from "./utils/theme"
 
 import ProtectedRoute from "./components/Auth/ProtectedRoute"
 import AppForm from "./components/SubmissionForm/AppForm"
@@ -10,32 +11,8 @@ import { AppsGallery } from "./components/AppCard"
 import Home from "./components/LandingPage/Home"
 import { NavBar, Header, Footer } from "./components/reusable"
 
-import { MuiThemeProvider, createMuiTheme } from "@material-ui/core"
+import { MuiThemeProvider } from "@material-ui/core"
 import "./App.scss"
-
-const theme = createMuiTheme({
-  typography: {
-    fontFamily: "Lato",
-  },
-  palette: {
-    primary: {
-      main: "#1a61b0",
-      contrastText: "#efeef3",
-    },
-    secondary: {
-      main: "#f32667",
-      contrastText: "#efeef3",
-    },
-    lambdaRed: {
-      main: "#bb1333",
-      contrastText: "#efeef3",
-    },
-    lambdaTeal: {
-      main: "#3ab5e5",
-      contrastText: "#efeef3",
-    },
-  },
-})
 
 const App = ({ loginUser }) => {
   const { loading } = useAuth0()
